TEAM-86 add salePrice

+++ src/main/java/com/tobias/productservice/inner/service/ProductService.java
@@ -9,7 +9,7 @@ public interface ProductService {
     void addProduct(RequestProduct product);
     Iterable<Product> getProductsAll();
     void deleteProduct(int itemId);
-    void setProductSale(int itemId,double saleRate);
+    void setProductSale(int itemId,int saleRate);
     List<Product> getProductsAllBySaleActive(boolean saleActive);
     void setProductCount(int id, int amount);
 }
